+++ models/__init__.py
@@ -1,2 +1,8 @@
 from .aroio import *
+from .audio import *
 from .auth import *
+from .convolver import *
+from .network import *
+from .streaming import *
+from .system import *
+from .webinterface import *
